refactor: Reduce form data argument chaining by making ReflectionUtil a service
- Renamed ReflectionUtil to ControllerHandler.
- Added the form data attribute to ControllerHandler.
- Removed form data from ObjectRequestParameter.
- Added a new argument to form data constructor to initialize the values with the provided ones from the request.
